feat(kernels): BF16 weight storage + inline matmul widening

+++ src/heads/flow/flow.h
@@ -5,6 +5,7 @@
 
 #include <array>
 #include <cstddef>
+#include <cstdint>
 #include <span>
 
 namespace fe {
@@ -49,12 +50,12 @@ private:
   }
 
   FlowConfig cfg_{};
-  const float* in_proj_{};                     // [hidden][action_dim]
-  const float* time_proj_{};                   // [hidden][time_dim]
-  const float* cond_proj_{};                   // [hidden][cond_dim]
-  std::array<const float*, kMaxMlp> layers_{}; // [hidden][hidden]
-  const float* out_proj_{};                    // [action_dim][hidden]
-  const float* freqs_{};                       // [time_dim/2] sinusoidal freqs
+  const uint16_t* in_proj_{};                     // [hidden][action_dim]
+  const uint16_t* time_proj_{};                   // [hidden][time_dim]
+  const uint16_t* cond_proj_{};                   // [hidden][cond_dim]
+  std::array<const uint16_t*, kMaxMlp> layers_{}; // [hidden][hidden]
+  const uint16_t* out_proj_{};                    // [action_dim][hidden]
+  const float* freqs_{};                          // [time_dim/2] computed at init
   Arena* scratch_{};
   bool ok_{false};
 };
